create Search folder and transfer logic
